Added a simple table

+++ models/dotcollection.py
@@ -12,7 +12,7 @@ class DotCollection:
     def __init__(self, xy=None, i=None, mask=None, fnos=None):
         self._xy = np.empty(shape=(0, 2), dtype=float) if xy is None else xy
         self._i = np.empty(shape=(0,), dtype=float) if i is None else i
-        self._fnos = np.zeros_like(self.i, dtype=float) if fnos is None else fnos
+        self._fnos = np.zeros_like(self.i, dtype=int) if fnos is None else fnos
         self.mask = np.ones_like(self.x, dtype=bool) if mask is None else mask
         assert (xy is None) == (i is None), "Both or neither of xy and i must be set"
         assert self._xy.shape[0] == self._i.shape[0], "xy must be of shape (N, 2) and m of shape (N,)"
